+++ src/test/java/org/folio/entlinks/integration/kafka/SpecificationChangedEventListenerTest.java
@@ -1,6 +1,9 @@
 package org.folio.entlinks.integration.kafka;
 
 import static java.util.UUID.randomUUID;
+import static org.folio.rspec.domain.dto.Family.MARC;
+import static org.folio.rspec.domain.dto.FamilyProfile.AUTHORITY;
+import static org.folio.rspec.domain.dto.FamilyProfile.BIBLIOGRAPHIC;
 import static org.folio.rspec.domain.dto.SpecificationUpdatedEvent.UpdateExtent.FULL;
 import static org.folio.rspec.domain.dto.SpecificationUpdatedEvent.UpdateExtent.PARTIAL;
 import static org.folio.support.base.TestConstants.TENANT_ID;
@@ -11,7 +14,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 
 import java.util.concurrent.Callable;
-import org.folio.entlinks.service.tenant.MarcSpecificationUpdateService;
+import org.folio.entlinks.integration.internal.MarcSpecificationUpdateService;
 import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.testing.type.UnitTest;
@@ -38,7 +41,17 @@ class SpecificationChangedEventListenerTest {
 
   @Test
   void handleEvent_positive_skipPartialUpdateEvents() {
-    var event = new SpecificationUpdatedEvent(randomUUID(), TENANT_ID, PARTIAL);
+    var event = new SpecificationUpdatedEvent(randomUUID(), TENANT_ID, MARC, BIBLIOGRAPHIC, PARTIAL);
+
+    listener.handleEvent(event);
+
+    verifyNoInteractions(executionService);
+    verifyNoInteractions(updateService);
+  }
+
+  @Test
+  void handleEvent_positive_skipFullAuthorityUpdateEvents() {
+    var event = new SpecificationUpdatedEvent(randomUUID(), TENANT_ID, MARC, AUTHORITY, FULL);
 
     listener.handleEvent(event);
 
@@ -48,7 +61,7 @@ class SpecificationChangedEventListenerTest {
 
   @Test
   void handleEvent_positive_callUpdateService() {
-    var event = new SpecificationUpdatedEvent(randomUUID(), TENANT_ID, FULL);
+    var event = new SpecificationUpdatedEvent(randomUUID(), TENANT_ID, MARC, BIBLIOGRAPHIC, FULL);
 
     listener.handleEvent(event);